fix: Split helpers module in two files, so they can be required by two files
license-checker-rseidelsohn and index both required the helpers module,
which did not work, as license-checker-rseidelsohn also requires index -
at least that's what I think was the culprit.
