+++ bin/license-checker-rseidelsohn
@@ -9,7 +9,7 @@ http://yuilibrary.com/license/
 const args = require('../lib/args');
 const exitProcessOrWarnIfNeeded = require('../lib/exitProcessOrWarnIfNeeded');
 const fs = require('fs');
-const helpers = require('../lib/helpers');
+const helpers = require('../lib/licenseCheckerHelpers');
 const licenseChecker = require('../lib/index');
 const mkdirp = require('mkdirp');
 const path = require('path');
