+++ lib/index.js
@@ -20,7 +20,7 @@ const createHash = require('crypto').createHash;
 
 const getLicenseTitle = require('./getLicenseTitle');
 const licenseFiles = require('./license-files');
-const helpers = require('../lib/helpers');
+const helpers = require('./indexHelpers');
 
 // Set up debug logging
 // https://www.npmjs.com/package/debug#stderr-vs-stdout
